Add AbstractPollCreate
